refactor: detach static component logic & simplify functions
* Put constants variables outside of the component such that it doesn't need to be re-computed every time the component is being rendered
* Replaced unnecessary function getters with nullish coalescing operator (`??`)
* Replaced unnecessary boolean declaration with ternaries
* Unified tournament time related functions to a single function that returns a tuple of start / end time
* removed component's static width requirement, allowing parent component to control it
